.prepareRasterBackedFile -- disallow overwrites if coming from new save
This is the fix for a long standing bug where attempts to recover raster from cache fail because the file backing the raster is gone. This was diagnosed: it is when 2 functions output same file, then one is deleted via clearCache. The other cache event is still there, but its file-backed component is gone.